\name{family.summary.cond}
\alias{family.summary.cond}
\title{
  Use family() on a ``summary.cond'' object
}
\description{
  This is a method for the function \code{family()} for objects 
  inheriting from class \code{summary.cond}.  See 
  \code{\link[stats]{family}} for the general behaviour of this 
  function.
}
\usage{
\method{family}{summary.cond}(object, \dots)
}
\arguments{
  \item{object}{
    any object from which a \code{family} object can be extracted.   
  }
  \item{\dots}{
    absorbs any additional argument.  
  }
}
\seealso{
  \code{\link[stats]{family}}
}
\keyword{classes} 
\keyword{methods}
